+++ jdbc/src/DBModule.kt
@@ -1,25 +1,16 @@
 package klite.jdbc
 
-import com.zaxxer.hikari.HikariConfig
-import com.zaxxer.hikari.HikariDataSource
-import klite.*
+import klite.Config
+import klite.Extension
+import klite.Server
 import klite.StatusCode.Companion.Conflict
+import klite.register
 import javax.sql.DataSource
 
 val Config.dbPoolMaxSize get() = optional("DB_POOL_SIZE")?.toInt() ?: ((optional("NUM_WORKERS")?.toInt() ?: 5) + (optional("JOB_WORKERS")?.toInt() ?: 5))
 
-class DBModule(val dataSource: DataSource): Extension {
-  constructor(configure: HikariConfig.() -> Unit = {}): this(HikariDataSource(HikariConfig().apply {
-    poolName = "app-db"
-    jdbcUrl = Config.required("DB_URL")
-    username = Config.optional("DB_USER")
-    password = Config.optional("DB_PASS")
-    minimumIdle = 1
-    maximumPoolSize = Config.dbPoolMaxSize
-    configure()
-    logger().info("Connecting to $jdbcUrl${username?.let { ", user: $username" }}")
-  }))
-
+/** See [HikariModule] if you want to use Hikari */
+open class DBModule(val dataSource: DataSource = PooledDataSource(ConfigDataSource())): Extension {
   override fun install(server: Server) = server.run {
     registry.register<DataSource>(dataSource)
     errors.on(AlreadyExistsException::class, Conflict)
@@ -27,7 +18,7 @@ class DBModule(val dataSource: DataSource): Extension {
   }
 }
 
-/** Call this before to add support for Heroku-like DATABASE_URL env variable */
+/** Use before DBModule to add support for Heroku-like DATABASE_URL env variable (non-jdbc) */
 fun initHerokuDB(suffix: String = "?sslmode=require") {
   val url = Config.optional("DATABASE_URL") ?: return
   val m = "postgres://(?<user>.+?):(?<password>.+?)@(?<hostportdb>.*)".toRegex().matchEntire(url)?.groups ?: return