+++ jobs/test/JobRunnerTest.kt
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test
 import java.time.LocalDate
 import java.time.LocalTime
 import java.util.concurrent.ScheduledExecutorService
-import java.util.concurrent.TimeUnit.MINUTES
+import java.util.concurrent.TimeUnit.MILLISECONDS
 import java.util.concurrent.TimeUnit.SECONDS
 import javax.sql.DataSource
 
@@ -44,19 +44,19 @@ class JobRunnerTest {
 
   @Test fun `scheduleDaily with delay`() {
     runner.scheduleDaily(job, delayMinutes = 10)
-    verify { executor.scheduleAtFixedRate(any(), 10, 24 * 60, MINUTES) }
+    verify { executor.scheduleAtFixedRate(any(), 10 * 60 * 1000, 86400000, MILLISECONDS) }
   }
 
   @Test fun `scheduleDaily at time`() {
     runner.scheduleDaily(job, LocalTime.of(6, 30), LocalTime.of(6, 45))
-    verify(exactly = 2) { executor.scheduleAtFixedRate(any(), match { it > 0 }, 24 * 60, MINUTES) }
+    verify(exactly = 2) { executor.scheduleAtFixedRate(any(), match { it > 0 }, 86400000, MILLISECONDS) }
   }
 
   @Test fun `scheduleMonthly today`() {
     val job = spyk(job)
     runner.scheduleMonthly(job, LocalDate.now().dayOfMonth, LocalTime.of(6, 30))
     val dailyRunner = slot<Runnable>()
-    verify { executor.scheduleAtFixedRate(capture(dailyRunner), match { it > 0 }, 24 * 60, MINUTES) }
+    verify { executor.scheduleAtFixedRate(capture(dailyRunner), match { it > 0 }, 86400000, MILLISECONDS) }
     dailyRunner.captured.run()
     coVerify { job.run() }
   }
@@ -65,7 +65,7 @@ class JobRunnerTest {
     val job = spyk(job)
     runner.scheduleMonthly(job, LocalDate.now().dayOfMonth + 1, LocalTime.of(6, 30))
     val dailyRunner = slot<Runnable>()
-    verify { executor.scheduleAtFixedRate(capture(dailyRunner), match { it > 0 }, 24 * 60, MINUTES) }
+    verify { executor.scheduleAtFixedRate(capture(dailyRunner), match { it > 0 }, 86400000, MILLISECONDS) }
     dailyRunner.captured.run()
     coVerify(exactly = 0) { job.run() }
   }